requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
distlib==0.3.1
Django==3.0.7
django-yearlessdate==1.3.1
djangorestframework==3.11.0
drf-yasg==1.17.1
filelock==3.0.12
flake8==3.8.3
idna==2.10
importlib-metadata==1.6.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.14.0
pytz==2019.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
tox==3.16.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.26
wcwidth==0.1.9
zipp==3.1.0
